Removed gl4es.h dependency in array.c
