+++ src/gl/array.c
@@ -2,9 +2,9 @@
 
 #include "debug.h"
 #include "enum_info.h"
-#include "state.h"
-#include "gl4es.h"
+#include "glcase.h"
 #include "light.h"
+#include "state.h"
 
 GLvoid *copy_gl_array(const GLvoid *src,
                       GLenum from, GLsizei width, GLsizei stride,
