+++ src/gl/fpe.c
@@ -5,6 +5,7 @@
 #include "debug.h"
 #include "enum_info.h"
 #include "fpe_shader.h"
+#include "glcase.h"
 #include "init.h"
 #include "loader.h"
 #include "matrix.h"
